feat: crud

+++ internal/adapters/api/user/handlers.go
@@ -1,20 +1,22 @@
 package user
 
 import (
-	"context"
 	api "dev-hack-backend/internal/adapters/api"
 	"dev-hack-backend/internal/service/user"
 	"dev-hack-backend/pkg/apperror"
 	"github.com/gin-gonic/gin"
+	"log"
 	"net/http"
 	"time"
 )
 
 const (
-	userURL      = "/:user_id"
-	userLogin    = "/login"
-	userRegister = "/register"
-	contextTTL   = time.Second * 15
+	userURL       = "/:user_id"
+	pathParamUser = "user_id"
+	userLogin     = "/login"
+	userRegister  = "/register"
+	//tokenRefresh  = "/refresh_token"
+	contextTTL = time.Second * 15
 )
 
 type handler struct {
@@ -31,6 +33,9 @@ func (h *handler) Register(router *gin.Engine) {
 	router.POST(userLogin, h.SignIn)
 	router.POST(userRegister, h.SignUp)
 
+	//auth := router.Group("/auth", h.tokenValid)
+	//auth.POST(tokenRefresh, h.RefreshToken)
+
 	userGroup := router.Group("/user", h.userIdentity)
 	userGroup.GET(userURL, h.GetUser)
 	userGroup.PATCH(userURL, h.PathUser)
@@ -38,7 +43,7 @@ func (h *handler) Register(router *gin.Engine) {
 }
 
 func (h *handler) SignIn(c *gin.Context) {
-	ctx, cancel := context.WithTimeout(c.Request.Context(), time.Second*15)
+	ctx, cancel := h.userService.ContextWithTimeout(c, contextTTL, pathParamUser)
 	defer cancel()
 
 	var userDTO SignInUserDTO
@@ -60,11 +65,11 @@ func (h *handler) SignIn(c *gin.Context) {
 		return
 	}
 
-	api.ResponseWithTokens(c, http.StatusCreated, aToken, rToken)
+	api.ResponseWithTokens(c, http.StatusCreated, currentUser.Id.Hex(), aToken, rToken)
 }
 
 func (h *handler) SignUp(c *gin.Context) {
-	ctx, cancel := h.userService.CreateContextWithTimeout(c.Request.Context(), contextTTL)
+	ctx, cancel := h.userService.ContextWithTimeout(c, contextTTL, pathParamUser)
 	defer cancel()
 
 	var userDTO CreateUserDTO
@@ -87,11 +92,19 @@ func (h *handler) SignUp(c *gin.Context) {
 		return
 	}
 
-	api.ResponseWithTokens(c, http.StatusCreated, aToken, rToken)
+	api.ResponseWithTokens(c, http.StatusCreated, currentUser.Id.Hex(), aToken, rToken)
 }
 
+//func (h *handler) RefreshToken(c *gin.Context) {
+//	ctx, cancel := h.userService.ContextWithTimeout(c, contextTTL, pathParamUser)
+//	defer cancel()
+//
+//	var token TokenDTO
+//
+//}
+
 func (h *handler) GetUser(c *gin.Context) {
-	ctx, cancel := h.userService.CreateContextWithTimeout(c.Request.Context(), contextTTL)
+	ctx, cancel := h.userService.ContextWithTimeout(c, contextTTL, pathParamUser)
 	defer cancel()
 
 	currentUser, err := h.userService.GetUser(ctx)
@@ -104,38 +117,48 @@ func (h *handler) GetUser(c *gin.Context) {
 }
 
 func (h *handler) PathUser(c *gin.Context) {
-	ctx, cancel := h.userService.CreateContextWithTimeout(c.Request.Context(), contextTTL)
+	ctx, cancel := h.userService.ContextWithTimeout(c, contextTTL, pathParamUser)
 	defer cancel()
 
-	var userDTO UpdateUserDTO
-	err := c.ShouldBindJSON(&userDTO)
+	currentUser, err := h.userService.GetUser(ctx)
+	if err != nil {
+		log.Fatal(err.Error())
+	}
+	err = c.ShouldBindJSON(&currentUser)
 	if err != nil {
 		api.NewResponse(c, http.StatusBadRequest, apperror.BadData.Error()+err.Error())
 		return
 	}
 
-	currentUser, err := userDTO.toUser(ctx)
-	if err != nil {
-		api.NewResponse(c, http.StatusInternalServerError, err.Error())
+	ok := h.userService.CompareID(ctx, c.Param(pathParamUser))
+	if !ok {
+		api.NewResponseStatus(c, http.StatusUnauthorized)
 		return
 	}
 
 	err = h.userService.UpdateUser(ctx, currentUser)
 	if err != nil {
 		api.NewResponse(c, http.StatusInternalServerError, err.Error())
+		return
 	}
 
-	api.ResponseUser(c, http.StatusOK, currentUser)
+	api.NewResponseStatus(c, http.StatusOK)
 }
 
 func (h *handler) DeleteUser(c *gin.Context) {
-	ctx, cancel := h.userService.CreateContextWithTimeout(c.Request.Context(), contextTTL)
+	ctx, cancel := h.userService.ContextWithTimeout(c, contextTTL, pathParamUser)
 	defer cancel()
+
+	ok := h.userService.CompareID(ctx, c.Param(pathParamUser))
+	if !ok {
+		api.NewResponseStatus(c, http.StatusUnauthorized)
+		return
+	}
 
 	err := h.userService.DeleteUser(ctx)
 	if err != nil {
 		api.NewResponse(c, http.StatusInternalServerError, err.Error())
 	}
 
-	api.NewResponseStatusCode(c, http.StatusNoContent)
+	api.NewResponseStatus(c, http.StatusNoContent)
 }
